requirements.txt
pandas==2.0.3
pillow
playsound==1.2.2
psutil==5.9.8
requests==2.31.0
torch==2.2.1
tqdm==4.66.2
ultralytics==8.1.19
imgaug==0.4.0